Modifying lots of readers to call MBReadUtilIface->update_adjacencies after creating a batch of elements and updating their connectivity. This ensures that, if AEntityFactory has already created vertex-element adjacencies, that those adjacencies will get created for the new elements/vertices.  Without this, some vertices have upward adjacencies and some do not.
Also updating mbcoupler_test to use new method for accessing
partitions/interfaces and calling parallel reader for multiple files.

Passes make check, except for parallel Tqdcfr test, which fails because of a check I inserted recently on number of partition sets vs. number of procs.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1923 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ WriteNCDF.cpp
@@ -1261,7 +1261,8 @@ MBErrorCode WriteNCDF::write_BCs(std::vector<NeumannSetData> &sidesets,
                                                   &(*begin_iter), 
                                                   1, &exodus_id);
       if (MB_FAILURE == result) {
-        mWriteIface->report_error("Problem getting exodus id for sideset element %lu", ID_FROM_HANDLE(*begin_iter));
+        mWriteIface->report_error("Problem getting exodus id for sideset element %lu", 
+                                  (long unsigned int) ID_FROM_HANDLE(*begin_iter));
         return result;
       }
